third_party/blink/web_tests/http/tests/websocket/reserved-bits-expected.txt

Test whether WebSocket rejects frames whose reserved bit is on.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".

Case compressed=false&bitNumber=1: Test started.
onopen() was called.
PASS onerror() was called
onclose() was called.
PASS closeEvent.wasClean is false
Case compressed=false&bitNumber=2: Test started.
onopen() was called.
PASS onerror() was called
onclose() was called.
PASS closeEvent.wasClean is false
Case compressed=false&bitNumber=3: Test started.
onopen() was called.
PASS onerror() was called
onclose() was called.
PASS closeEvent.wasClean is false
Case compressed=true&bitNumber=2: Test started.
onopen() was called.
PASS onerror() was called
onclose() was called.
PASS closeEvent.wasClean is false
Case compressed=true&bitNumber=3: Test started.
onopen() was called.
PASS onerror() was called
onclose() was called.
PASS closeEvent.wasClean is false
PASS successfullyParsed is true

TEST COMPLETE
